+++ apps/site/app/page.tsx
@@ -1,0 +1,5 @@
+import { OfficialSite } from "../src/OfficialSite";
+
+export default function Home() {
+  return <OfficialSite />;
+}